Maybe got no duplicate posts working?
I implemented a get operation in the bot function, such that it gets the bot's last post text. If the bot is trying to post the same thing it did last time, return a dummy 37 value instead of posting the post again. I turned the 5-10 minute faucet back on, we'll see if this works :-) If so, yippee I've got a basic Walt Ruff bot working!!!!!