Long posts parse into reply threads
Long posts that need more than one post to be posted on bsky now post their multiple posts as a thread under the original, instead of creating separate posts. Changed the bot.post signature to enable a bool arg value that tells the post whether or not to post this post as an original post, or as a reply to the previous post. Root URI and CID are kept as public variables to the class.